# Avoid an unnecessary allocation in recording rule evaluation