--- a/rules/recording.go
+++ b/rules/recording.go
@@ -89,7 +89,8 @@ func (rule *RecordingRule) Eval(ctx context.Context, evalDelay time.Duration, ts
 			lb.Set(l.Name, l.Value)
 		}
 
-		sample.Metric = lb.Labels(nil)
+		labelSetSize := len(sample.Metric) + len(rule.labels) + 1 // additional 1 for rule name
+		sample.Metric = lb.Labels(make(labels.Labels, 0, labelSetSize))
 	}
 
 	// Check that the rule does not produce identical metrics after applying