@charleskorn 2023-01-04
I've just realised that this change won't work with https://github.com/prometheus/prometheus/pull/10991... so I'm not sure if there's value in introducing this if we're expecting to bring in that improvement soon.