@pracucci 2023-01-04
If we decide to move forward, I would like to see this PR in `prometheus/prometheus`, and not in our fork. Then we'll just update the upstream here.

> I've just realised that this change won't work with https://github.com/prometheus/prometheus/pull/10991... so I'm not sure if there's value in introducing this if we're expecting to bring in that improvement soon.

I think we should move forward with whatever optimization we find in the current implementation, which can be rediscussed once 10991 will be close to be merged (it's still a draft and not yet approved in Prometheus).