@charleskorn 2023-01-05
Closing in favour of https://github.com/prometheus/prometheus/pull/11812.